+++ lib/Components/FilterBar/FilterBar.jsx
@@ -1,4 +1,5 @@
 import ElementFilter from './ElementFilter';
+import WrittenOnlyToggle from './WrittenOnlyToggle';
 
 import './FilterBar.scss';
 
@@ -7,6 +8,7 @@ export default function FilterBar({ variables }) {
     <div className="bio-vo-filter-bar">
       <ElementFilter variables={ variables } />
       <div className="bio-vo-filter-bar-actions">
+        <WrittenOnlyToggle />
       </div>
     </div>
   );